Use Model Binding

+++ event-management/app/Http/Controllers/Api/AttendeeController.php
@@ -8,11 +8,13 @@ use App\Http\Traits\CanLoadRelationships;
 use App\Models\Attendee;
 use App\Models\Event;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class AttendeeController extends Controller
 {
     use CanLoadRelationships;
     private array $relations = ['user'];
+
     public function index(Event $event)
     {
         $attendees = $this->loadRelationships(
@@ -43,8 +45,9 @@ class AttendeeController extends Controller
         //
     }
 
-    public function destroy(string $event, Attendee $attendee)
+    public function destroy(Event $event, Attendee $attendee)
     {
+        Gate::authorize('delete-attendee', [$event, $attendee]);
         $attendee->delete();
         return response(status: 204);
     }
